Added JSONArraySerializer and JSONObjectSerializer
In the course of working with JSON, sometimes you just want to be able
to pass JSON around (say, in Android Activities) without having to
convert to and from strings all the time. These decorators still
convert to and from string, but they implement Serializable so that you
don't have to play games. Just stick your data in there and get on with
your life.